requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
frozenlist==1.4.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.10
multidict==6.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
yarl==1.9.2
